requirements.txt
streamlit
opencv-python-headless
numpy
img2pdf
Pillow